Remove the NSS_LIBS and KRB5_LIBS variables from sssd.spec
Due to the way RPM processes the %configure macro, these variables
were not actually being passed down to recursive configure
invocations. In other words, they were useless.

Futhermore, in more recent Fedora versions (13+), some of the
dependencies have moved from -lnss to -lnspr4. As a result, it is
safer to rely on the complete output of 'pkg-config nss --libs'
instead of restricting to -lnss. The downside to this is that it
may result in linking unnecessarily against other NSS components
such as libsmime3 and libplc4 (among others). However, since these
are already dependencies of libnss itself, there should be no risk
of them being unavailable on the platform when installed.