Fix ini_config unit test
When running 'make distcheck', the entire source directory is set
to read-only, to ensure that the build process only has write
access to $builddir. As a result, this was causing the unit test
for file mode to fail, since the file it was testing resides in
the $srcdir.

This patch guarantees that the test file has the correct
permissions prior to running the access test.