Resend SIGTERM if child doesn't terminate
There is a race condition where if we send a SIGTERM before the
kernel has scheduled the child, it may be lost, and the child will
not terminate and will leave the monitor hung in wait().

This patch alters this behavior so that we will send the SIGTERM
again every 10ms and check the wait() in a nonblocking manner.