Allow fallback to SSSD domain
For backwards-compatibility with older versions of the SSSD (such
as 1.2.x), we need to be able to have our DNS SRV record lookup be
capable of falling back to using the SSSD domain name as the DNS
discovery domain.

This patch modifies our DNS lookups so that they behave as
follows:

If dns_discovery_domain is specified, it is considered
authoritative. No other discovery domains will be attempted.

If dns_discovery_domain is not specified, we first attempt to look
up the SRV records using the domain portion of the machine's
hostname. If this returns "NOTFOUND", we will try performing an
SRV record query using the SSSD domain name as the DNS discovery
domain.

https://fedorahosted.org/sssd/ticket/754