Don't double-sanitize member DNs
After asking the cache for the list of member DNs for groups
during an initgroups request, we were passing it through the
sanitization function. Since this had already been done before
they were saved to the cache, this meant that it was corrupting
the results. It is safe to pass the returned DN directly into the
sysdb_group_dn_name() function.