Add additional indexing for sysdb
Adds an index for dataExpireTimestamp
This is used for determining which users need to be removed during
the cleanup task. If enumeration is enabled (or huge numbers of
users have been cached), the cleanup task runs very slowly due to
the non-indexed search.

Also adds an index for ONELEVEL lookups, to speed up situations
where we would need to request all entries under a particular node
in the LDB.