Fix uninitialized value error in ipa_get_id_options()
Previously, we were only constructing the basedn variable if the
ldap_search_base was not specified (which is unlikely to be in use
when using the IPA provier). However, if it did happen,
constrcuction of the compat search base for netgroups would be
using an uninitialized value.

Fixes https://fedorahosted.org/sssd/ticket/806