Fix one unlikely case of failure in sdap_id_op module
There can be an unlikely scenario when the first part of
sdap_id_op_connect_done works fine and there is no need to mark backend
offline. But right after the check, the memory allocation can fail in
which case the backend needs to be marked offline along with disabled
reconnecting.