Keep deref controls until the whole request is finished
https://fedorahosted.org/sssd/ticket/989

John Hodrien found out that when paging is used while dereferencing an
entry, sssd_be may segfault on the second page.

This was because paging returned the control to sdap_generic_search
multiple times but sssd was freeing dereference control after the first
search invocation. The subsequend sdap searched accessed memory that was
already freed.