Unbreak ./configure
./configure at least from 1.5.13 is failing on Ubuntu Oneiric.

The node ``Conditionals'' of automake manual states:

Note that you must arrange for _every_ `AM_CONDITIONAL' to be invoked
every time `configure' is run.  If `AM_CONDITIONAL' is run conditionally
(e.g., in a shell `if' statement), then the result will confuse `automake'.

So the trick is to run AM_CONDITIONAL unconditionally.