Add tool to convert debug levels
Older versions of SSSD (1.5 and earlier) would take a debug_level
value set in the [sssd] section as authoritative for all other
sections where not explicitly overridden. We changed this so that
all sections need to set it if they want debug logs set.
This script can be run to make the new version continue to produce
the same logs as the old versions did, by explicitly adding
debug_level to all domains and services that did not have it set
already.

Bugzilla https://bugzilla.redhat.com/show_bug.cgi?id=753763