Fix building manpages in parallel build dirs
Commit 2ebb3cf3d7062860343c3db1b53173c9c6547038 introduced an issue
when building manpages in parallel build dirs. It unnecessarily
modified the XML_DOC definition to restrict it to basing its
contents on the man_MANS variable. Unfortunately, this means that
when we entered the recursion in all-local we would be effectively
turning XML_DOC into $(srcdir)/$(lang)/manpage.N.xml when we really
only care that the source language XML document exists.

This patch reverts that portion of the change.

https://fedorahosted.org/sssd/ticket/1293