Ghost members - NSS responder changes
Since there are two attributes storing information about user
memberships of the group we have to include both of them in results.

This will apply only for objects that have ghost members (i.e. they
contain the SYSDB_GHOST attribute). If an object has this attribute,
values of this attribute are not projected to the memberuid attribute.