sudo ldap provider: notify responder when an expired rule has been deleted
When an expired rule is not present on the server server during specific rule
refresh, the provider will notify the sudo responder that it has been deleted.

Because there is a high probability that some other rules were deleted from
the server as well, we want to remove them from sysdb as soon as possible.

Once the responder is notified, it will schedule an out of band full refresh.
This is issued by responder, because we already have a mechanism that
prohibits creation of similar request (i.e. once the OOB full refresh is
scheduled, there won't be another).

The notification is done by returning:
DP error = DP_ERR_OK, error = ENOENT