Modify priority evaluation in SELinux user maps
The functionality now is following:

When rule is being matched, its priority is determined as a combination
of user and host specificity (host taking preference).

After the rule is matched in provider, only its host priority is stored
in sysdb for later usage.

When rules are matched in the responder, their user priority is
determined. After that their host priority is retrieved directly from
sysdb and sum of both priorities is user to determine whether to use
that rule or not. If more rules have the same priority, the order given
in IPA config is used.

https://fedorahosted.org/sssd/ticket/1360
https://fedorahosted.org/sssd/ticket/1395