remove left over principal selection
https://fedorahosted.org/sssd/ticket/1303

Domain start up was taking too long when there are many principals
in a kerberos keytab. We were looking up in the keytab two times.

The first time we try to select a proper principal and remember it.
The second call happens almost right after the first one and
it is just a check if the principal exists in the keytab, without
any output information other than success/failure. It is
probably a left over from https://fedorahosted.org/sssd/ticket/781.

This patch removes the second call.