do not fail if POLLHUP occurs while reading data
This cause troubles when we send data to a pipe and close the
file descriptor before data is read. The pipe is still readable,
but POLLHUP is detected and we fail to read them.

For example, this may cause a user beeing unable to log in.

Now if POLLHUP appears, we read the pipe and then close it on
the client side too.