do not create pid file twice
If a provider is terminated and the monitor tries to restart it,
it goes again through mark_service_as_started() which
will try to create pid file again because number of running
services didn't change.

Because the pid file cannot be created twice, it will not return
EOK and the whole SSSD is terminated.