Fix segfault when ID-mapping an entry without a SID
If there was no SID attribute, then we would have detected it by
checking the number of values of an element. We would however happily
return EOK in that case and save garbage into the sid_str.

This was causing segfault when the entry was supposed to be ID-mapped by
had no SID.