Make sub-domains case-insensitive
Currently the only type of supported sub-domains are AD domains which
are not case-sensitive. To make it easier for Windows user we make
sub-domains case-insensitive as well which allows to write the username
in any case at the login prompt.

If support for other types of sub-domains is added it might be necessary
to set the case-sensitive flag based on the domain type.